+++ CHANGELOG.md
@@ -22,7 +22,8 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
     - `Py42ResponseError`
     - `Py42UserAlreadyAddedError`
     - `Py42UserNotInLegalHoldError`
-    - `Py42LegalHoldNotFoundOrPermissionDeniedError`
+    - `Py42LegalHoldNotFoundError`
+    - `Py42PermissionDeniedError`
     - `Py42UserDoesNotExistError`
     - `Py42InvalidRuleTypeError`
 
@@ -70,7 +71,8 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 - `py42.detectionlists.departing_employee.add()` now raises `Py42UserAlreadyAddedError` when the user is already on the list.
 - `py42.detectionlists.high_risk_employee.add()` now raises `Py42UserAlreadyAddedError` when the user already on the list.
 - `py42.legalhold.add_to_matter()` now raises `Py42UserAlreadyAddedError` when the user is already on the matter.
-- `py42.legalhold.get_matter_by_uid()` now raises `Py42LegalHoldNotFoundOrPermissionDeniedError` when the matter is unavailable.
+- `py42.legalhold.get_matter_by_uid()` now raises `Py42LegalHoldNotFoundError` on bad requests.
+- `py42.legalhold.get_matter_by_uid()` now raises `Py42LegalHoldPermissionDeniedError` when the user does not have access.
 - `py42.users.get_by_username()` now raises `Py42UserDoesNotExistError` when the user is not found.
 - `py42.alerts.rules.add_user()` now raises `Py42InvalidRuleTypeError` when trying to add a user to a system rule.
 - `py42.alerts.rules.remove_user()` now raises `Py42InvalidRuleTypeError` when trying to remove user from a system rule.